* Reorganise code * Use timeout during http client initialisation and stop using it later in context
